requirements.txt
aiohttp==3.8.1
PyCryptodome==3.12.0
requests==2.26.0